bugfix: _extract_traceback
`index` was counting lines, not just of the traceback but of leading
outputs as well (which are collected in `all_else`).
Wether a line contained the location data or the corresponding source
code was determined by checking if `index` was odd or even.
This worked if there was an odd number of lines before the stacktrace
but if there was no output before the stacktrace it skipped the first
line of location info, interpreted the corresponding source code line
as location info and the location info of the next entry as it's source
code.
Depending on the number of commas in the corresponding source code line
this problem was more or less obvious.
I have solved this by toggling a boolean variable called `location_info`
instead of using `index`.

The last entry of the stack trace is skipped because the 2nd for loop
in _extract_traceback iterates over `entries[:-2]` where `entries[-1]`
is the error message and `entries[-2]` is the last entry of the stack
trace.