changed comment
This comment seems wrong to me. This block is executed if capture
is False. capture is set to True when starting to parse the stack
trace and never reset to False. Therefore it is executed for each
line *before* the stack trace.
Also, there should be nothing after the stack trace.